refactor: make exceptions private
