+++ src/iminuit/_exceptions.py
