+++ tests/test_deprecated.py
@@ -1,5 +1,5 @@
 from iminuit._deprecated import deprecated, deprecated_parameter
-from iminuit.exceptions import VisibleDeprecationWarning
+from iminuit._exceptions import VisibleDeprecationWarning
 import pytest
